Fixed premissions

+++ src/main/java/App/controllers/department/DepartmentController.java
@@ -1,5 +1,6 @@
 package App.controllers.department;
 
+import App.controllers.enums.AuthResponse;
 import App.controllers.enums.DepartmentResponse;
 import App.entity.Department;
 import App.models.department.DepartmentRegisterModel;
@@ -9,6 +10,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.util.List;
 import java.util.Optional;
@@ -49,34 +51,42 @@ public class DepartmentController {
     }
 
     @PostMapping
-    public ResponseEntity createDepartment(@Valid @RequestBody DepartmentRegisterModel departmentName) {
-        if(this.departmentService.findDepartmentByDepartmentName(departmentName.getDepartmentName()).isPresent()) {
-            return new ResponseEntity<>(DepartmentResponse.ALREADY_EXISTS.toString(), HttpStatus.BAD_REQUEST);
+    public ResponseEntity createDepartment(@Valid @RequestBody DepartmentRegisterModel departmentName, HttpServletRequest request) {
+        if(request.isUserInRole("ADMIN")) {
+            if (this.departmentService.findDepartmentByDepartmentName(departmentName.getDepartmentName()).isPresent()) {
+                return new ResponseEntity<>(DepartmentResponse.ALREADY_EXISTS.toString(), HttpStatus.BAD_REQUEST);
+            }
+
+            try {
+                Department department = new Department(departmentName.getDepartmentName());
+                this.departmentService.createOrUpdate(department);
+
+                return ok(department);
+            } catch (Exception ex) {
+                return new ResponseEntity<>(DepartmentResponse.ERROR.toString(), HttpStatus.BAD_REQUEST);
+            }
         }
 
-        try {
-            Department department = new Department(departmentName.getDepartmentName());
-            this.departmentService.createOrUpdate(department);
-
-            return ok(department);
-        } catch(Exception ex) {
-            return new ResponseEntity<>(DepartmentResponse.ERROR.toString(), HttpStatus.BAD_REQUEST);
-        }
+        return new ResponseEntity<>(AuthResponse.NO_PREMISSIONS.toString(), HttpStatus.BAD_REQUEST);
     }
 
     @DeleteMapping(value = "{id}")
-    public ResponseEntity deleteDepartment(@Valid @PathVariable String id) {
-        Optional<Department> department = this.departmentService.findById(UUID.fromString(id));
+    public ResponseEntity deleteDepartment(@Valid @PathVariable String id, HttpServletRequest request) {
+        if(request.isUserInRole("ADMIN")) {
+            Optional<Department> department = this.departmentService.findById(UUID.fromString(id));
 
-        if(department.isPresent()) {
-            try {
-                this.departmentService.delete(department.get());
-                return new ResponseEntity<>(DepartmentResponse.SUCCESSFULLY_DELETED.toString(), HttpStatus.OK);
-            } catch(Exception ex) {
-                return new ResponseEntity<>(DepartmentResponse.ERROR.toString(), HttpStatus.BAD_REQUEST);
+            if (department.isPresent()) {
+                try {
+                    this.departmentService.delete(department.get());
+                    return new ResponseEntity<>(DepartmentResponse.SUCCESSFULLY_DELETED.toString(), HttpStatus.OK);
+                } catch (Exception ex) {
+                    return new ResponseEntity<>(DepartmentResponse.ERROR.toString(), HttpStatus.BAD_REQUEST);
+                }
+            } else {
+                return new ResponseEntity<>(DepartmentResponse.NO_DEPARTMENT.toString(), HttpStatus.BAD_REQUEST);
             }
-        } else {
-            return new ResponseEntity<>(DepartmentResponse.NO_DEPARTMENT.toString(), HttpStatus.BAD_REQUEST);
         }
+
+        return new ResponseEntity<>(AuthResponse.NO_PREMISSIONS.toString(), HttpStatus.BAD_REQUEST);
     }
 }
